refactor: convert virtualdevices command to yargs
