+++ src/commands/index.ts
@@ -17,6 +17,7 @@ import locationsDeleteCommand from './locations/delete.js'
 import locationsUpdateCommand from './locations/update.js'
 import organizationsCommand from './organizations.js'
 import schemaCreateCommand from './schema/create.js'
+import virtualdevicesCommand from './virtualdevices.js'
 
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -38,4 +39,5 @@ export const commands: CommandModule<object, any>[] = [
 	locationsUpdateCommand,
 	organizationsCommand,
 	schemaCreateCommand,
+	virtualdevicesCommand,
 ]
